tests/baseline/protocol-1.2/inspector/console/console-format-expected.txt

CONSOLE MESSAGE: line 27: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 28: %o
CONSOLE MESSAGE: line 29: %O
CONSOLE MESSAGE: line 30: Test for zero "%f" in formatter
CONSOLE MESSAGE: line 31: %% self-escape1
CONSOLE MESSAGE: line 32: %%s self-escape2
CONSOLE MESSAGE: line 33: %%ss self-escape3
CONSOLE MESSAGE: line 34: %%s%s%%s self-escape4
CONSOLE MESSAGE: line 35: %%%%% self-escape5
CONSOLE MESSAGE: line 36: %%%s self-escape6
CONSOLE MESSAGE: line 12: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 13: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 12: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 13: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 12: test
CONSOLE MESSAGE: line 13: test
CONSOLE MESSAGE: line 12: test named "test"
CONSOLE MESSAGE: line 13: test named "test"
CONSOLE MESSAGE: line 12: Error
CONSOLE MESSAGE: line 13: Error
CONSOLE MESSAGE: line 12: Error: my error message
CONSOLE MESSAGE: line 13: Error: my error message
CONSOLE MESSAGE: line 12: Error: my multiline
error message
CONSOLE MESSAGE: line 13: Error: my multiline
error message
CONSOLE MESSAGE: line 12: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 13: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 12: function () { return 1; }
CONSOLE MESSAGE: line 13: function () { return 1; }
CONSOLE MESSAGE: line 12: function () {
        return 2;
    }
CONSOLE MESSAGE: line 13: function () {
        return 2;
    }
CONSOLE MESSAGE: line 12: 0.12
CONSOLE MESSAGE: line 13: 0.12
CONSOLE MESSAGE: line 12: http://webkit.org/
CONSOLE MESSAGE: line 13: http://webkit.org/
CONSOLE MESSAGE: line 12: null
CONSOLE MESSAGE: line 13: 
CONSOLE MESSAGE: line 12: undefined
CONSOLE MESSAGE: line 13: 
CONSOLE MESSAGE: line 12: [object Attr]
CONSOLE MESSAGE: line 13: [object Attr]
CONSOLE MESSAGE: line 12: [object Attr]
CONSOLE MESSAGE: line 13: [object Attr]
CONSOLE MESSAGE: line 12: [object Attr]
CONSOLE MESSAGE: line 13: [object Attr]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: NaN
CONSOLE MESSAGE: line 13: NaN
CONSOLE MESSAGE: line 12: Infinity
CONSOLE MESSAGE: line 13: Infinity
CONSOLE MESSAGE: line 12: -Infinity
CONSOLE MESSAGE: line 13: -Infinity
CONSOLE MESSAGE: line 12: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 13: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: function () {}
CONSOLE MESSAGE: line 13: function () {}
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: [object SVGSVGElement]
CONSOLE MESSAGE: line 13: [object SVGSVGElement]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: 0
CONSOLE MESSAGE: line 13: 0
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: function Object() { [native code] }
CONSOLE MESSAGE: line 13: function Object() { [native code] }
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
CONSOLE MESSAGE: line 13: function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
CONSOLE MESSAGE: line 12: 42
CONSOLE MESSAGE: line 13: 42
CONSOLE MESSAGE: line 12: abc
CONSOLE MESSAGE: line 13: abc
CONSOLE MESSAGE: line 12: [object Uint16Array]
CONSOLE MESSAGE: line 13: [object Uint16Array]
CONSOLE MESSAGE: line 12: [object Text]
CONSOLE MESSAGE: line 13: [object Text]
CONSOLE MESSAGE: line 12: [object DOMException]
CONSOLE MESSAGE: line 13: [object DOMException]
CONSOLE MESSAGE: line 12: [object Uint8Array]
CONSOLE MESSAGE: line 13: [object Uint8Array]
CONSOLE MESSAGE: line 12: [object Uint8Array]
CONSOLE MESSAGE: line 13: [object Uint8Array]
CONSOLE MESSAGE: line 12: [object Uint8Array]
CONSOLE MESSAGE: line 13: [object Uint8Array]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: 1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1
CONSOLE MESSAGE: line 13: 1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1
CONSOLE MESSAGE: line 12: test
CONSOLE MESSAGE: line 13: test
Tests that console logging dumps proper messages.

 console-format.html:22 Array[10]
console-format.html:23 Array[10]
console-format.html:24 Array[10]
console-format.html:25 Test for zero "0" in formatter
console-format.html:26 % self-escape1 dummy
console-format.html:27 %s self-escape2 dummy
console-format.html:28 %ss self-escape3 dummy
console-format.html:29 %sdummy%s self-escape4
console-format.html:30 %%% self-escape5 dummy
console-format.html:31 %dummy self-escape6
console-format.html:7 /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:8 [/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\…?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i]
globals[0]
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:7 /foo\\bar\sbaz/i
console-format.html:8 [/foo\\bar\sbaz/i]
globals[1]
/foo\\bar\sbaz/i
console-format.html:7 test
console-format.html:8 ["test"]
globals[2]
"test"
console-format.html:7 test named "test"
console-format.html:8 ["test named "test""]
globals[3]
"test named "test""
console-format.html:7 Error
console-format.html:8 [Error
]
globals[4]
Error
console-format.html:7 Error: my error message
console-format.html:8 [Error: my error message
]
globals[5]
Error: my error message
console-format.html:7 Error: my multiline
error message
console-format.html:8 [Error: my multiline
error message
]
globals[6]
Error: my multiline
error message
console-format.html:7 
    <p id="p">Tests that console logging dumps proper messages.</p>
console-format.html:8 [p#p]
globals[7]
    <p id="p">Tests that console logging dumps proper messages.</p>
console-format.html:7 function () { return 1; }
console-format.html:8 [function]
globals[8]
function () { return 1; }
console-format.html:7 function () {
        return 2;
    }
console-format.html:8 [function]
globals[9]
function () {
        return 2;
    }
console-format.html:7 0.12
console-format.html:8 [0.12]
globals[10]
0.12
console-format.html:7 http://webkit.org/
console-format.html:8 ["http://webkit.org/"]
globals[11]
"http://webkit.org/"
console-format.html:7 null
console-format.html:8 [null]
globals[12]
null
console-format.html:7 undefined
console-format.html:8 [undefined]
globals[13]
undefined
console-format.html:7 
    attr=""
console-format.html:8 [attr]
globals[14]
    attr=""
console-format.html:7 
    attr="value"
console-format.html:8 [attr]
globals[15]
    attr="value"
console-format.html:7 
    id="x"
console-format.html:8 [id]
globals[16]
    id="x"
console-format.html:7 Object {}
console-format.html:8 [Object]
globals[17]
Object {}
console-format.html:7 NaN
console-format.html:8 [NaN]
globals[18]
NaN
console-format.html:7 Infinity
console-format.html:8 [Infinity]
globals[19]
Infinity
console-format.html:7 -Infinity
console-format.html:8 [-Infinity]
globals[20]
-Infinity
console-format.html:7 ["test", "test2", undefined × 2, "test4", undefined × 5, foo: Object]
console-format.html:8 [Array[10]]
globals[21]
["test", "test2", undefined × 2, "test4", undefined × 5, foo: Object]
console-format.html:7 Object {}
console-format.html:8 [Object]
globals[22]
Object {}
console-format.html:7 [function]
console-format.html:8 [Array[1]]
globals[23]
[function]
console-format.html:7 Object {bar: "bar"}
console-format.html:8 [Object]
globals[24]
Object {bar: "bar"}
console-format.html:7 
    <svg id="svg-node"></svg>
console-format.html:8 [svg#svg-node]
globals[25]
    <svg id="svg-node"></svg>
console-format.html:7 Object {enumerableProp: 4, __underscoreEnumerableProp__: 5, __underscoreNonEnumerableProp: 2, abc: 3}
console-format.html:8 [Object]
globals[26]
Object {enumerableProp: 4, __underscoreEnumerableProp__: 5, __underscoreNonEnumerableProp: 2, abc: 3}
console-format.html:7 -0
console-format.html:8 [-0]
globals[27]
-0
console-format.html:7 Object {}
console-format.html:8 [Object]
globals[28]
Object {}
console-format.html:7 function Object() { [native code] }
console-format.html:8 [function]
globals[29]
function Object() { [native code] }
console-format.html:7 Object {}
console-format.html:8 [Object]
globals[30]
Object {}
console-format.html:7 function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
console-format.html:8 [function]
globals[31]
function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
console-format.html:7 Number {[[PrimitiveValue]]: 42}
console-format.html:8 [Number]
globals[32]
Number {[[PrimitiveValue]]: 42}
console-format.html:7 String {0: "a", 1: "b", 2: "c", length: 3, [[PrimitiveValue]]: "abc"}
console-format.html:8 [String]
globals[33]
String {0: "a", 1: "b", 2: "c", length: 3, [[PrimitiveValue]]: "abc"}
console-format.html:7 Uint16Array[3] [1, 2, 3]
console-format.html:8 [Uint16Array[3]]
globals[34]
Uint16Array[3] [1, 2, 3]
console-format.html:7 #text
console-format.html:8 [text]
globals[35]
#text
console-format.html:7 DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
console-format.html:8 [DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of th…]
globals[36]
DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
console-format.html:7 Uint8Array[1] [3]
console-format.html:8 [Uint8Array[1]]
globals[37]
Uint8Array[1] [3]
console-format.html:7 Uint8Array[400] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
console-format.html:8 [Uint8Array[400]]
globals[38]
Uint8Array[400] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
console-format.html:7 Uint8Array[400000000] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
console-format.html:8 [Uint8Array[400000000]]
globals[39]
Uint8Array[400000000] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
console-format.html:7 namespace.longSubNamespace.x.className {}
console-format.html:8 [n…e.l…e.x.className]
globals[40]
namespace.longSubNamespace.x.className {}
console-format.html:7 [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…]
console-format.html:8 [Array[200]]
globals[41]
[1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…]
console-format.html:7 ["test"]
console-format.html:8 [Array[1]]
globals[42]
["test"]
Expanded all messages
console-format.html:22 Array[10]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:23 Array[10]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:24 Array[10]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:25 Test for zero "0" in formatter
console-format.html:26 % self-escape1 dummy
console-format.html:27 %s self-escape2 dummy
console-format.html:28 %ss self-escape3 dummy
console-format.html:29 %sdummy%s self-escape4
console-format.html:30 %%% self-escape5 dummy
console-format.html:31 %dummy self-escape6
console-format.html:7 /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:8 [/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\…?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i]
    0: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
    length: 1
    __proto__: Array[0]
globals[0]
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:7 /foo\\bar\sbaz/i
console-format.html:8 [/foo\\bar\sbaz/i]
    0: /foo\\bar\sbaz/i
    length: 1
    __proto__: Array[0]
globals[1]
/foo\\bar\sbaz/i
console-format.html:7 test
console-format.html:8 ["test"]
    0: "test"
    length: 1
    __proto__: Array[0]
globals[2]
"test"
console-format.html:7 test named "test"
console-format.html:8 ["test named "test""]
    0: "test named "test""
    length: 1
    __proto__: Array[0]
globals[3]
"test named "test""
console-format.html:7 Error
console-format.html:8 [Error
]
    0: Error
    length: 1
    __proto__: Array[0]
globals[4]
Error
console-format.html:7 Error: my error message
console-format.html:8 [Error: my error message
]
    0: Error: my error message
    length: 1
    __proto__: Array[0]
globals[5]
Error: my error message
console-format.html:7 Error: my multiline
error message
console-format.html:8 [Error: my multiline
error message
]
    0: Error: my multiline
error message
    length: 1
    __proto__: Array[0]
globals[6]
Error: my multiline
error message
console-format.html:7 
    <p id="p">Tests that console logging dumps proper messages.</p>
console-format.html:8 [p#p]
    0: p#p
    length: 1
    __proto__: Array[0]
globals[7]
    <p id="p">Tests that console logging dumps proper messages.</p>
console-format.html:7 function () { return 1; }
console-format.html:8 [function]
    0: function ()
    length: 1
    __proto__: Array[0]
globals[8]
function () { return 1; }
console-format.html:7 function () {
        return 2;
    }
console-format.html:8 [function]
    0: function ()
    length: 1
    __proto__: Array[0]
globals[9]
function () {
        return 2;
    }
console-format.html:7 0.12
console-format.html:8 [0.12]
    0: 0.12
    length: 1
    __proto__: Array[0]
globals[10]
0.12
console-format.html:7 http://webkit.org/
console-format.html:8 ["http://webkit.org/"]
    0: "http://webkit.org/"
    length: 1
    __proto__: Array[0]
globals[11]
"http://webkit.org/"
console-format.html:7 null
console-format.html:8 [null]
    0: null
    length: 1
    __proto__: Array[0]
globals[12]
null
console-format.html:7 undefined
console-format.html:8 [undefined]
    0: undefined
    length: 1
    __proto__: Array[0]
globals[13]
undefined
console-format.html:7 
    attr=""
console-format.html:8 [attr]
    0: attr
    length: 1
    __proto__: Array[0]
globals[14]
    attr=""
console-format.html:7 
    attr="value"
console-format.html:8 [attr]
    0: attr
    length: 1
    __proto__: Array[0]
globals[15]
    attr="value"
console-format.html:7 
    id="x"
console-format.html:8 [id]
    0: id
    length: 1
    __proto__: Array[0]
globals[16]
    id="x"
console-format.html:7 Object {}
    length: (...)
    get length: function length()
    __proto__: Object
console-format.html:8 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[17]
Object {}
    length: (...)
    get length: function length()
    __proto__: Object
console-format.html:7 NaN
console-format.html:8 [NaN]
    0: NaN
    length: 1
    __proto__: Array[0]
globals[18]
NaN
console-format.html:7 Infinity
console-format.html:8 [Infinity]
    0: Infinity
    length: 1
    __proto__: Array[0]
globals[19]
Infinity
console-format.html:7 -Infinity
console-format.html:8 [-Infinity]
    0: -Infinity
    length: 1
    __proto__: Array[0]
globals[20]
-Infinity
console-format.html:7 ["test", "test2", undefined × 2, "test4", undefined × 5, foo: Object]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:8 [Array[10]]
    0: Array[10]
    length: 1
    __proto__: Array[0]
globals[21]
["test", "test2", undefined × 2, "test4", undefined × 5, foo: Object]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:7 Object {}
    __proto__: Object
console-format.html:8 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[22]
Object {}
    __proto__: Object
console-format.html:7 [function]
    0: function ()
    length: 1
    __proto__: Array[0]
console-format.html:8 [Array[1]]
    0: Array[1]
    length: 1
    __proto__: Array[0]
globals[23]
[function]
    0: function ()
    length: 1
    __proto__: Array[0]
console-format.html:7 Object {bar: "bar"}
    bar: "bar"
    __proto__: Object
console-format.html:8 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[24]
Object {bar: "bar"}
    bar: "bar"
    __proto__: Object
console-format.html:7 
    <svg id="svg-node"></svg>
console-format.html:8 [svg#svg-node]
    0: svg#svg-node
    length: 1
    __proto__: Array[0]
globals[25]
    <svg id="svg-node"></svg>
console-format.html:7 Object {enumerableProp: 4, __underscoreEnumerableProp__: 5, __underscoreNonEnumerableProp: 2, abc: 3}
    enumerableProp: 4
    __underscoreEnumerableProp__: 5
    abc: 3
    bar: (...)
    getFoo: function ()
    __underscoreNonEnumerableProp: 2
    get bar: function ()
    set bar: function (x)
    __proto__: Object
console-format.html:8 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[26]
Object {enumerableProp: 4, __underscoreEnumerableProp__: 5, __underscoreNonEnumerableProp: 2, abc: 3}
    enumerableProp: 4
    __underscoreEnumerableProp__: 5
    abc: 3
    bar: (...)
    getFoo: function ()
    __underscoreNonEnumerableProp: 2
    get bar: function ()
    set bar: function (x)
    __proto__: Object
console-format.html:7 -0
console-format.html:8 [-0]
    0: -0
    length: 1
    __proto__: Array[0]
globals[27]
-0
console-format.html:7 Object {}
    No Properties
console-format.html:8 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[28]
Object {}
    No Properties
console-format.html:7 function Object() { [native code] }
console-format.html:8 [function]
    0: function Object()
    length: 1
    __proto__: Array[0]
globals[29]
function Object() { [native code] }
console-format.html:7 Object {}
    constructor: function Object()
    hasOwnProperty: function hasOwnProperty()
    isPrototypeOf: function isPrototypeOf()
    propertyIsEnumerable: function propertyIsEnumerable()
    toLocaleString: function toLocaleString()
    toString: function toString()
    valueOf: function valueOf()
    __defineGetter__: function __defineGetter__()
    __defineSetter__: function __defineSetter__()
    __lookupGetter__: function __lookupGetter__()
    __lookupSetter__: function __lookupSetter__()
    get __proto__: function __proto__()
    set __proto__: function __proto__()
console-format.html:8 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[30]
Object {}
    constructor: function Object()
    hasOwnProperty: function hasOwnProperty()
    isPrototypeOf: function isPrototypeOf()
    propertyIsEnumerable: function propertyIsEnumerable()
    toLocaleString: function toLocaleString()
    toString: function toString()
    valueOf: function valueOf()
    __defineGetter__: function __defineGetter__()
    __defineSetter__: function __defineSetter__()
    __lookupGetter__: function __lookupGetter__()
    __lookupSetter__: function __lookupSetter__()
    get __proto__: function __proto__()
    set __proto__: function __proto__()
console-format.html:7 function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
console-format.html:8 [function]
    0: function ( /**/ foo/**/, /*/**/bar,     /**/baz)
    length: 1
    __proto__: Array[0]
globals[31]
function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
console-format.html:7 Number {[[PrimitiveValue]]: 42}
    __proto__: Number
    [[PrimitiveValue]]: 42
console-format.html:8 [Number]
    0: Number
    length: 1
    __proto__: Array[0]
globals[32]
Number {[[PrimitiveValue]]: 42}
    __proto__: Number
    [[PrimitiveValue]]: 42
console-format.html:7 String {0: "a", 1: "b", 2: "c", length: 3, [[PrimitiveValue]]: "abc"}
    0: "a"
    1: "b"
    2: "c"
    length: 3
    __proto__: String
    [[PrimitiveValue]]: "abc"
console-format.html:8 [String]
    0: String
    length: 1
    __proto__: Array[0]
globals[33]
String {0: "a", 1: "b", 2: "c", length: 3, [[PrimitiveValue]]: "abc"}
    0: "a"
    1: "b"
    2: "c"
    length: 3
    __proto__: String
    [[PrimitiveValue]]: "abc"
console-format.html:7 Uint16Array[3] [1, 2, 3]
    0: 1
    1: 2
    2: 3
    buffer: (...)
    byteLength: (...)
    byteOffset: (...)
    length: (...)
    Symbol(Symbol.toStringTag): (...)
    __proto__: TypedArray
console-format.html:8 [Uint16Array[3]]
    0: Uint16Array[3]
    length: 1
    __proto__: Array[0]
globals[34]
Uint16Array[3] [1, 2, 3]
    0: 1
    1: 2
    2: 3
    buffer: (...)
    byteLength: (...)
    byteOffset: (...)
    length: (...)
    Symbol(Symbol.toStringTag): (...)
    __proto__: TypedArray
console-format.html:7 #text
console-format.html:8 [text]
    0: text
    length: 1
    __proto__: Array[0]
globals[35]
#text
console-format.html:7 DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
console-format.html:8 [DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of th…]
    0: DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
    length: 1
    __proto__: Array[0]
globals[36]
DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
console-format.html:7 Uint8Array[1] [3]
    0: 3
    buffer: (...)
    byteLength: (...)
    byteOffset: (...)
    length: (...)
    Symbol(Symbol.toStringTag): (...)
    __proto__: TypedArray
console-format.html:8 [Uint8Array[1]]
    0: Uint8Array[1]
    length: 1
    __proto__: Array[0]
globals[37]
Uint8Array[1] [3]
    0: 3
    buffer: (...)
    byteLength: (...)
    byteOffset: (...)
    length: (...)
    Symbol(Symbol.toStringTag): (...)
    __proto__: TypedArray
console-format.html:7 Uint8Array[400] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
    [0 … 99]
    [100 … 199]
    [200 … 299]
    [300 … 399]
    foo: "bar"
    __proto__: TypedArray
console-format.html:8 [Uint8Array[400]]
    0: Uint8Array[400]
    length: 1
    __proto__: Array[0]
globals[38]
Uint8Array[400] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
    [0 … 99]
    [100 … 199]
    [200 … 299]
    [300 … 399]
    foo: "bar"
    __proto__: TypedArray
console-format.html:7 Uint8Array[400000000] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
    [0 … 99999999]
    [100000000 … 199999999]
    [200000000 … 299999999]
    [300000000 … 399999999]
    __proto__: TypedArray
console-format.html:8 [Uint8Array[400000000]]
    0: Uint8Array[400000000]
    length: 1
    __proto__: Array[0]
globals[39]
Uint8Array[400000000] [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
    [0 … 99999999]
    [100000000 … 199999999]
    [200000000 … 299999999]
    [300000000 … 399999999]
    __proto__: TypedArray
console-format.html:7 namespace.longSubNamespace.x.className {}
    __proto__: Object
console-format.html:8 [n…e.l…e.x.className]
    0: namespace.longSubNamespace.x.className
    length: 1
    __proto__: Array[0]
globals[40]
namespace.longSubNamespace.x.className {}
    __proto__: Object
console-format.html:7 [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…]
    [0 … 99]
    [100 … 199]
    length: 200
    __proto__: Array[0]
console-format.html:8 [Array[200]]
    0: Array[200]
    length: 1
    __proto__: Array[0]
globals[41]
[1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…]
    [0 … 99]
    [100 … 199]
    length: 200
    __proto__: Array[0]
console-format.html:7 ["test"]
    0: "test"
    length: 1
    __proto__: Array[0]
console-format.html:8 [Array[1]]
    0: Array[1]
    length: 1
    __proto__: Array[0]
globals[42]
["test"]
    0: "test"
    length: 1
    __proto__: Array[0]
